# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This documentation describes MiniStream API",
        "title": "MiniStream API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/job/{jobuuid}": {
            "get": {
                "description": "Get the job description and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Get a job",
                "operationId": "job-get",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid.UUID",
                        "description": "Some job UUID",
                        "name": "jobuuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Delete a job",
                "operationId": "job-delete",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid.UUID",
                        "description": "Some job UUID",
                        "name": "jobuuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/login": {
            "get": {
                "description": "Logs user into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logs user into the system",
                "operationId": "user-login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ACCESS-KEY-ID",
                        "name": "ACCESS-KEY-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "SECRET-ACCESS-KEY",
                        "name": "SECRET-ACCESS-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/stream.LoginUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apierror.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/utils/pbkdf2": {
            "post": {
                "description": "Generate hash from password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utils"
                ],
                "summary": "Generate hash from password",
                "operationId": "utils-pbkdf2",
                "parameters": [
                    {
                        "format": "Pbkdf2Payload",
                        "description": "Pbkdf2Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.Pbkdf2Payload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierror.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apierror.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "details": {
                    "description": "application-level error details that best describes the error, for debugging",
                    "type": "string"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "streamUUID": {
                    "description": "stream uuid",
                    "type": "string"
                },
                "validationErrors": {
                    "description": "list of errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apierror.ValidationError"
                    }
                }
            }
        },
        "apierror.ValidationError": {
            "type": "object",
            "properties": {
                "failedField": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "stream.LoginUserResponse": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "web.Pbkdf2Payload": {
            "type": "object",
            "required": [
                "digest",
                "password",
                "salt"
            ],
            "properties": {
                "digest": {
                    "type": "string",
                    "example": "sha256"
                },
                "iterations": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "password": {
                    "type": "string",
                    "example": "thisismysecretpassword"
                },
                "salt": {
                    "type": "string",
                    "example": "thisisarandomsalt"
                }
            }
        }
    }
}